Move encryption code to separate CipherUtil object
